Refactor transform tests
Refactored tests away from snapshots to provide more explicit
information about what a given test case is looking for in the output.
Previous snapshots would fail if the transform didn't match the expected
output, but with little pointers on exactly what the expected outcome
should be.

Normally this is something that should be easy to spot with the naked
eye but this becomes harder where things like spacing are concerned. A
good example of this being "allows leading and trailing whitespace" test
case but with little information about exactly where the whitespace
should be. The current solution using `expect.to.include` probably also
leaves a lot to be desired but hopefully it (in conjunction with the
renaming of the test case) provides a little more clarity around exactly
what is being looked for.

Additionally the usage of "should" was removed in test case names. This results
in test case names that state facts rather than aspirational outcomes (i.e:
"this code WILL" rather than "this code SHOULD")