Docker fixes (#95)
* Switch to docker compose plugin

Usage of the `docker-compose` command is deprecated and instead it is
generally recommended that people use `docker compose` (note the lack of
hyphen) going forward. To that end, this commit replaces the invoation
of `docker-compose` with `docker compose`.

More info: https://stackoverflow.com/a/66526176

* Replace docker host DNS alias with service name

As a convenience docker desktop provides a DNS alias
(host.docker.internal) that can be used to connect to services on the
host machine from within containers. However, this would fail to work for anyone
not docker desktop.

Therefore, using service names here is probably a better option.