Remove edge case for sentinel
This code is (at this time) impossible to test since it relies on
"sentinel" being available in `PRODUCT_CONFIG` which is currently
commented out.

The only way to write tests for this would be to mock out the
`productVersions.mjs` module. However, it probably makes sense (for the
time being at least) to remove this code entirely to avoid having
untested code lying around. It also removes the temptation to continue
supporting this edge case in the future if/when we migrate sentinel.