Add tests for GET /[productSlug]/[version]/[...docsPath] API route (#103)
* Sanity check productSlug before usage

Implementation of the first test case highlighted a case where requesting a
nonexistent product would cause errors to be thrown as the code was
trying to index the PRODUCT_CONFIG object using a product name that
doesn't exist.

* Move shared test fixtures to beforeEach block

The URL object, the Request factory function and the console.error stub
can all move to the beforeEach block to avoid having to repeat them in
each test. This helps remove clutter from tests and provide better
clarity around what the test is actually testing for

* Return a 404 for nonexistent versions

* Move params type to API function

Moving the params type to the API handler function and then importing
from there helps ensure consistency between the test code and the
production code and also helps provide guidance when calling the
function in test code

* Mock out contentVersions for better test isolation

This helps ensure that the tests aren't relying on code that isn't part
of the unit under test (in this case, the API route handler)

* Add test for missing content

Add test to ensure that 404 is returned if content is missing from the filesystem

* Add test to return 404 for invalid markdown

* Inspect arguments passed to console.error

Since the API route returns the same response (404: Not Found) for all
errors, it's helpful to have some additional code in there to make sure
that the test is failing for the expected reason

Therefore, some additional assertions were added to look at the
arguments passed to `console.error` to make sure that the implementation
code was erroring where it was supposed to

* Add test to assert that the requested docs are returned

* Remove edge case for sentinel

This code is (at this time) impossible to test since it relies on
"sentinel" being available in `PRODUCT_CONFIG` which is currently
commented out.

The only way to write tests for this would be to mock out the
`productVersions.mjs` module. However, it probably makes sense (for the
time being at least) to remove this code entirely to avoid having
untested code lying around. It also removes the temptation to continue
supporting this edge case in the future if/when we migrate sentinel.

* Types cleanup

Removed a mapped type that was no longer needed, and added some
docblocks to the exported type in the route handler to make things a
little clearer

* Simplify `mockRequest` function for tests

There was some confusion around how the URL parameters actually worked
with this handler. There was initial confusion about the apparent
disconnect between what looked like URLs with dynamic segments, and the
ability to have named request variables (implying GET parameters) in the
handler.

I'd forgotten how Next does (or can do) routing using the filesystem to
configure
routing (https://nextjs.org/docs/pages/building-your-application/routing/api-routes#dynamic-api-routes)
but essentially, the magic of mapping the URL segments to variables
was being done by Next under the hood based on folder structure.

In order to cut down on complexity, the `mockRequest` function was
updated to (somewhat) do this itself so that test cases didn't have to
worry about consistency between the URL path passed to the `mockRequest`
function and the params