Update test files
In attempting to debug an issue with a file path that was seemingly not
correct, it turned out that mocking out the filesystem using memfs was
also messing with the ability to import things in tests.

Therefore, importing and mocking out individual bits of `fs` seemed to
yield a better result than using and `memfs`

An additional check was aded to check for a file's existence before
proceeding as fatal errors were being thrown when trying to process (and
by extension JSON decode) missing files. This was resulting in
`undefined` being passed to `JSON.parse()` which was in turn halting the
execution.

Additionally, both functions were tested together as mocking each one
out to test the other provide very difficult (half a day was spent
trying to get this to work with `vitest`), likely because they're in the
same file and `vi.mock` works by monkeypatching the `import` system in
node (this is just a guess though).

Either way, since it could be argued that one is just an implementation
detail of the other and largely a black box, they were both just tested
together as one big black box