Lint and test only main and PRs
Instead of running tests against every branch (even ones without active
PRs), we instead decided that we want to only run tests against PRs.
Branches _without_ a PR should be treated as WIP or draft. This is (in
part) because we are charged for GHA execution so cutting down on
unnecessary GHA execution gives us an easy cost saving.

More discussion on this can be found here:
https://github.com/hashicorp/web-unified-docs/pull/113#discussion_r1940084606

Co-authored-by: Heat Hamilton <heat.hamilton@hashicorp.com>