Fix hardcoded test values in tests
The tests for getProductVersion relied on hard-coded test values that
were easily outdated. This meant that as soon as `versionMetadata.json`
was updated, the values of what was considered "latest" changed and
broke the tests.

Therefore, this commit updated those tests to hopefully be a tad more
dynamic (and less prone to breakage) going forward