Fix filesystme mocks in tests
The implementation code was previously using usage of `import` to bring
in the version metadata file (rather than fs.*). This meant that `memfs`
probably wasn't doing much to mock the file contents in the tests.

In addition, the implementation code was trying to import
`versionMetadata.json` while the tests were trying to mock calls for
`version-metadata.json`.

The fix was to swap out `import` for fs.readFileSync and then use
`memfs` to create an in-memory filesystem to read from.