Resolve test, lint & typing issues (#136)
* Lint and test only main and PRs

Instead of running tests against every branch (even ones without active
PRs), we instead decided that we want to only run tests against PRs.
Branches _without_ a PR should be treated as WIP or draft. This is (in
part) because we are charged for GHA execution so cutting down on
unnecessary GHA execution gives us an easy cost saving.

More discussion on this can be found here:
https://github.com/hashicorp/web-unified-docs/pull/113#discussion_r1940084606

Co-authored-by: Heat Hamilton <heat.hamilton@hashicorp.com>

* Update eslint config

Increases the number of files that eslint ignores including content and
repo configuration files (i.e: vitest config)

* Fix hardcoded test values in tests

The tests for getProductVersion relied on hard-coded test values that
were easily outdated. This meant that as soon as `versionMetadata.json`
was updated, the values of what was considered "latest" changed and
broke the tests.

Therefore, this commit updated those tests to hopefully be a tad more
dynamic (and less prone to breakage) going forward

* eslint

---------

Co-authored-by: Heat Hamilton <heat.hamilton@hashicorp.com>
Co-authored-by: Ruben Nic <RubenSandwich@users.noreply.github.com>