Compare char count using fromJson function
This compares the char count to the action variable numerically (integer to
integer) rather than lexicographically. In this case, the report char count was
'1209122' which is _technically_ greater than 65536(the configured max), but
since the values were being compared lexicographically, and the string '1209122'
begins with a 1, this expression was evaluating to '1209122' >= '65536'(since
'1209122' begins with a 1).