Use docsPathsAllVersions to store document ID
In order to link two documents across versions, some kind of mapping is needed
to link document IDs to filepaths on disk.

Since this file already contains all the info we need for later retrieval (path
and version in particular) adding the ID in as well seemed like a sensible next
step rather than re-inventing the wheel (in the form of yet *another* generated
metadata file).

The approach of syncronously reading each file during the metadata
building stage was chosen as this is where the schema of docsPathsAllVersions
is effectively set as objects are pushed onto the `apiPaths` array.

Moving this logic elsewhere would introduce an additional build step and
would complicate things by splitting the "source of truth" for the
schema format into two places.