Use mock filesystem in tests
Update the other two tests to use memfs to mock the filesystem (rather than
relying on files to be present in the REAL filesytem). Also cleaned up
the tests a little to to use variables for commonly re-used values (like
product slug, version and file path). This helps make the tests less
dependent on magic values (and therefore less likely to break if one of
those values is changed).