Implement API route handler
Use the document ID to look up all available documents in the metadata file to
find other files with the same ID.

The product slug was intentionally omitted from this API endpoint
implementation. This is because the ID is (or should be) globally unique
across all documents, therefore the product can be (if required)
determined by looking at where a specific document with a specific ID
was located.

However, the current implementation has the potential to slow down as the
metadata file increases in size, so a potential enhancement might be to do two
searches: the first to look up the first occurance of an ID, then the
second to scope the remainder of the search under that product. This has
not been tested or profiled though, so some investigation would be required to
see if this actually would be any faster...