Updating example plugin name in Oracle DB examples (#1142)
As of Vault v1.9.0 (per commit
30a483088d8e3ed9f659f6072c9ee5f3a6dae90e), Vault's UI does not support
custom plugin names.  The OracleDB plugin's officially supported name is
`vault-plugin-database-oracle'.  Using another name will cause Vault's
UI to throw a warning to the user, and redirect them to the CLI/API.

This change just updates the plugin name used in the examples for the
Oracle Database Secret Engine doc to use the officially support name, so
as to reduce confusion for customers who may encounter this warning in
the UI.

Co-authored-by: Sarah Chavis <62406755+schavis@users.noreply.github.com>