Nomad: recommendations for singleton deployments (#1473)
Many users have a requirement to run exactly one instance of a given allocation
because it requires exclusive access to some cluster-wide resource, which we'll
refer to here as a "singleton allocation". This is challenging to implement, so
this document is intended to describe an accepted design to publish as a
how-to/tutorial.

Co-authored-by: Aimee Ukasick <aimee.ukasick@hashicorp.com>